+++ src/service/SchemaService/SchemaService.ts
@@ -1,23 +1,6 @@
 import { AxiosResponse } from 'axios';
-import { APIClient, APIMapping } from '../http';
+import { APIClient, APIMapping } from '../../http';
 
-export namespace SchemaServiceTypes {
-    export interface IntegrationFormData {
-        customCss?: string;
-        customJs?: string;
-        customRedirectUrl?: string;
-        customSubmitName?: string;
-        integrationFields?: Record<string, unknown>[];
-    }
-    export interface Integration {
-        formData: IntegrationFormData;
-        id: string;
-        label: string;
-        schemaId: string;
-        crudUris?: Record<string, unknown>;
-        dataUris?: Record<string, unknown>;
-    }
-}
 export class SchemaService extends APIClient {
     constructor() {
         super(APIMapping.schemaService);
