+++ src/service/SchemaService/index.ts
@@ -1,0 +1,2 @@
+export { default as SchemaService } from './SchemaService';
+export * from './SchemaService.Types';
